requirements.txt
streamlit
openai-whisper
transformers
torch
moviepy
pydub
ffmpeg-python